Added Ralink USB WLAN drivers for RT2870, RT3070, RT3370, RT35xx, RT3573, RT5370, RT5572
Because the 2,4GHz band here in Berlin becomes overloaded, I decided to switch to 5GHz.
For this I selected an USB-Stick with an Ralink chip because of the known good kernel
support. The devie I got has an RT5372 chip. The kernel also ha a driver for this chip,
but - shocking - market as 'not functional'. And yes, the driver which comes with the
kernel does not work at all except printing out multiple error messages on the console.

So I took the latest driver source directly from Ralinks website, compiled and tested
it. Here is the working result :-)